11 as a player i want to be able to destroy obstacles so that i can more easily shoot my target (#56)
Created cactus prefab. it responds to the Bullet tag colliding and will
remove parts of the cactus based on the relative x position. if it
reaches the final segment and is hit, the object is destroyed.

Because the end segments exist off the ground. If the cactus is hit from
one side only until the left sprite == the right sprite, it'll delete
automatically.